+++ src/filter.cpp
@@ -94,7 +94,7 @@ PNG::Result PNG::AdaptiveFiltering::UnfilterPixels(size_t width, size_t height, 
                 size_t pIndex = pixelStart >> 3; // pixelStart / 8
                 size_t pShift = 8 - (pixelStart & 7) - pixelBits; // pixelStart % 8
                 size_t byte = (unf[y][pIndex] >> pShift) & pixelMask;
-                out[y][x] = byte;
+                out[y][x] = (uint8_t)byte;
             }
         }
     }
